[CL-2438] Move idea_custom_fields out of module and further generalization of form builder
