+++ front/app/components/FormBuilder/edit/utils.ts
@@ -1,7 +1,5 @@
 import { questionDNDType } from 'components/FormBuilder/components/FormFields';
 import { IFlatCustomField } from 'services/formCustomFields';
-import { ParticipationMethod } from 'services/participationContexts';
-import { ideationConfig, nativeSurveyConfig } from '../configurations';
 
 const reorder = <ListType>(
   list: ListType[],
@@ -33,13 +31,6 @@ export type DragAndDropResult = {
   reason: string;
   mode: string;
   destination: DragOrDroValues | null;
-};
-
-export const getBuilderConfig = (participationMethod: ParticipationMethod) => {
-  if (participationMethod === 'ideation') {
-    return ideationConfig;
-  }
-  return nativeSurveyConfig;
 };
 
 const getFlatPageStructure = (
