+++ front/app/containers/Admin/projects/project/nativeSurvey/messages.tsx
@@ -1,0 +1,8 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  survey: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.BuilderTitle',
+    defaultMessage: 'Survey',
+  },
+});
