--- front/app/modules/commercial/idea_custom_fields/admin/components/messages.ts
@@ -1,8 +1,0 @@
-import { defineMessages } from 'react-intl';
-
-export default defineMessages({
-  inputFormTab: {
-    id: 'app.containers.AdminPage.ProjectEdit.inputFormTab',
-    defaultMessage: 'Input form',
-  },
-});
